refactor the thinking comp
